requirements.txt
requests>=2.32.0
reportlab>=4.2.0
Pillow>=10.4.0
streamlit>=1.37.0
pytest>=8.3.0
easyocr>=1.7.1
pytesseract>=0.3.10
